+++ src/Recyclarr.Core/Config/Parsing/PostProcessing/ConfigDeprecationPostProcessor.cs
@@ -1,11 +1,15 @@
 using System.Diagnostics.CodeAnalysis;
 using Recyclarr.Config.Parsing.PostProcessing.Deprecations;
 using Recyclarr.Logging;
+using Recyclarr.Sync.Events;
 using Serilog.Context;
 
 namespace Recyclarr.Config.Parsing.PostProcessing;
 
-public class ConfigDeprecationPostProcessor(ConfigDeprecations deprecations) : IConfigPostProcessor
+public class ConfigDeprecationPostProcessor(
+    ConfigDeprecations deprecations,
+    ISyncScopeFactory scopeFactory
+) : IConfigPostProcessor
 {
     [SuppressMessage("ReSharper", "WithExpressionModifiesAllMembers")]
     public RootConfigYaml Process(RootConfigYaml config)
@@ -27,6 +31,7 @@ public class ConfigDeprecationPostProcessor(ConfigDeprecations deprecations) : I
         where T : ServiceConfigYaml
     {
         using var logScope = LogContext.PushProperty(LogProperty.Scope, instanceName);
+        using var instanceScope = scopeFactory.SetInstance(instanceName);
         return deprecations.CheckAndTransform(yaml);
     }
 }