+++ src/Recyclarr.Core/CoreAutofacModule.cs
@@ -271,6 +271,10 @@ public class CoreAutofacModule : Module
     private static void RegisterSyncEvents(ContainerBuilder builder)
     {
         builder.RegisterType<SyncEventStorage>().SingleInstance();
-        builder.RegisterType<SyncEventCollector>().As<ISyncEventCollector>().SingleInstance();
+        builder
+            .RegisterType<SyncEventCollector>()
+            .As<ISyncScopeFactory>()
+            .As<ISyncEventPublisher>()
+            .SingleInstance();
     }
 }